@@ -555,9 +555,8 @@ extension BFFAPIService: TargetType {
     //API 基础地址
     var baseURL: URL {
   //      let baseUrl = "http://192.168.201.21:6380"  //本地
- //       let baseUrl = "http://erf.gazo.net.cn:6380" //测试
- //     let baseUrl = "https://api.gazolife.cn" //生产
-        let baseUrl = "http://erf.gazo.net.cn:6380"
+   //     let baseUrl = "http://erf.gazo.net.cn:6380" //测试
+      let baseUrl = "https://api.gazolife.cn" //生产
         return URL(string: baseUrl)!
     }
     
